@@ -3,21 +3,25 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.14",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/googlemaps": "^3.30.16",
-    "core-js": "^2.5.4",
+    "core-js": "^3.6.2",
     "git-gui": "0.0.7",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.4",
     "sweetalert2": "^7.28.2",
-    "zone.js": "^0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "^0.10.2"
   },
   "scripts": {
     "ng": "ng",
@@ -28,24 +32,27 @@
     "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.2",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "^0.803.21",
+    "@angular/cli": "~8.3.21",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^7.2.15",
+    "@types/jasmine": "^3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "@types/node": "^13.1.4",
+    "codelyzer": "^5.2.1",
+    "jasmine-core": "^3.5.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "karma-jasmine": "^3.0.1",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "protractor": "^5.4.2",
+    "rxjs-tslint": "^0.1.7",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "tslint": "^5.20.1",
+    "typescript": "~3.4.5",
+    "webpack": "^4.41.5"
   }
 }
