@@ -1,7 +1,7 @@
 // This file can be replaced during build by using the `fileReplacements` array.
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
-
+import 'zone.js/dist/zone-error';
 export const environment = {
   production: false,
   blockchainServiceUrl: 'http://localhost:8080/api/auth',
